Implement Emarsys::Email.export_responses and Emarsys::Export.data
It is now possible to create an export via API, and to download it once it is
done.

When creating an export, contact fields need to be given as an array. To support
converting the identifiers to field IDs, Emarsys::ParamsConverter was adapted to
deal with arrays, too.

The export download is a special endpoint that does not return JSON, but
text/csv if the call is successful (JSON otherwise). To support this,
Emarsys::Response needed to handle non-JSON responses, too.