fixes a segmentation fault problem when executed in forked environments
e.g. when running in background workers like `resque`

the actual bug is in Ruby (https://bugs.ruby-lang.org/issues/13753), but to also
allow older 2.2.x releases to make use of this gem i swapped the call to
`Random::DEFAULT.bytes` with `SecureRandom.random_bytes` which is
not affected by this problem but should work in nearly similar ways.